+++ internal/compiler/lookup.rs
@@ -831,6 +831,8 @@ impl LookupObject for ColorFunctions {
         let mut f = |n, e: Expression| f(n, e.into());
         None.or_else(|| f("rgb", BuiltinMacroReference(BuiltinMacroFunction::Rgb, t.clone())))
             .or_else(|| f("rgba", BuiltinMacroReference(BuiltinMacroFunction::Rgb, t.clone())))
+            .or_else(|| f("hsv", BuiltinMacroReference(BuiltinMacroFunction::Hsv, t.clone())))
+            .or_else(|| f("hsva", BuiltinMacroReference(BuiltinMacroFunction::Hsv, t.clone())))
     }
 }
 
@@ -996,7 +998,13 @@ impl<'a> LookupObject for ColorExpression<'a> {
                 )),
             })
         };
-        None.or_else(|| f("brighter", member_function(BuiltinFunction::ColorBrighter)))
+        None.or_else(|| f("red", member_function(BuiltinFunction::ColorRed)))
+            .or_else(|| f("green", member_function(BuiltinFunction::ColorGreen)))
+            .or_else(|| f("blue", member_function(BuiltinFunction::ColorBlue)))
+            .or_else(|| f("hue", member_function(BuiltinFunction::ColorHue)))
+            .or_else(|| f("saturation", member_function(BuiltinFunction::ColorSaturation)))
+            .or_else(|| f("brightness", member_function(BuiltinFunction::ColorBrightness)))
+            .or_else(|| f("brighter", member_function(BuiltinFunction::ColorBrighter)))
             .or_else(|| f("darker", member_function(BuiltinFunction::ColorDarker)))
             .or_else(|| f("transparentize", member_function(BuiltinFunction::ColorTransparentize)))
             .or_else(|| f("with-alpha", member_function(BuiltinFunction::ColorWithAlpha)))