+++ CHANGELOG.md
@@ -22,6 +22,10 @@ All notable changes to this project are documented in this file.
  - Image: Added support for 9 slice scaling
  - Image: Added `horizontal-` and `vertical-tiling`
  - Flickable: Added `flicked` callback
+ - Cupertino TabWidget: Tweaked visual appearance
+ - Slint: Expose `.red()`, `.green()`, `.blue()` methods on `color`
+ - Slint: Expose `.hue()`, `.saturation()`, `.brightness()` methods on `color`
+ - Slint: Add `hsv` and `hsva` as method to create colors
 
 ### Widgets